Varie prove per inserire i float numbers.

+++ project/database/migrations/2021_02_01_144147_create_peripherals_table.php
@@ -18,8 +18,8 @@ class CreatePeripheralsTable extends Migration
             
             $table -> string('name', 50);
             $table -> string('model', 100) -> unique();
-            $table -> smallInteger('price');
-            $table -> integer('expenditure');
+            $table -> float('price');
+            $table -> float('expenditure');
 
             $table -> timestamps();
         });
